Add keyvault service
